2017.8.2上架

+++ 立寻/MyFile/1.Classes/3.Sections/4.Main/2.MainSection/3.Issue/3.Sections/2.IssueInfo/4.PickUpSomeThing/1.Controller/PickUpSomeThingVC.m
@@ -513,7 +513,7 @@
         }else{
             count --;
         }
-        self.titleNumberLabel.text = [NSString stringWithFormat:@"%ld/20",count];
+        self.titleNumberLabel.text = [NSString stringWithFormat:@"%ld/20",(long)count];
     }
     return true;
 }
